requirements.txt
psycopg2-binary
pandas
sqlalchemy
validate-docbr
